resolved merge issues
