+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RobotSystem.java
@@ -3,9 +3,7 @@ package org.firstinspires.ftc.teamcode;
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.arcrobotics.ftclib.geometry.Pose2d;
-import com.arcrobotics.ftclib.geometry.Rotation2d;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
@@ -29,11 +27,10 @@ public class RobotSystem {
         this.telemetry = new MultipleTelemetry(opMode.telemetry, FtcDashboard.getInstance().getTelemetry());
         HardwareRobot hardwareRobot = new HardwareRobot(hardwareMap);
         this.cv = new CVSubsystem(
-                hardwareRobot.cameraName,
+                hardwareRobot.limelight,
                 initialPose.getHeading(),
                 isRedTeam,
-                telemetry
-        );
+                telemetry);
         this.localization = new LocalizationSubsystem(
                 initialPose,
                 hardwareRobot.leftOdometer,
