+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/autons/AutoBlueObservation.java
@@ -11,9 +11,11 @@ import org.firstinspires.ftc.teamcode.RobotSystem;
 import org.firstinspires.ftc.teamcode.opmodes.calibration.Drawing;
 import org.firstinspires.ftc.teamcode.synchropather.systems.__util__.Synchronizer;
 import org.firstinspires.ftc.teamcode.synchropather.systems.__util__.TimeSpan;
+import org.firstinspires.ftc.teamcode.synchropather.systems.claw.ClawConstants;
 import org.firstinspires.ftc.teamcode.synchropather.systems.claw.ClawPlan;
 import org.firstinspires.ftc.teamcode.synchropather.systems.claw.ClawState;
 import org.firstinspires.ftc.teamcode.synchropather.systems.claw.movements.LinearClaw;
+import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.ElbowConstants;
 import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.ElbowPlan;
 import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.ElbowState;
 import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.movements.LinearElbow;
@@ -26,23 +28,22 @@ import org.firstinspires.ftc.teamcode.synchropather.systems.rotation.movements.L
 import org.firstinspires.ftc.teamcode.synchropather.systems.translation.TranslationPlan;
 import org.firstinspires.ftc.teamcode.synchropather.systems.translation.TranslationState;
 import org.firstinspires.ftc.teamcode.synchropather.systems.translation.movements.CRSplineTranslation;
-import org.firstinspires.ftc.teamcode.synchropather.systems.translation.movements.LinearTranslation;
 
-@Autonomous(name="Auto Blue Human Specimen", group="Autons")
-public class AutoBlueHumanSpecimen extends LinearOpMode{
+@Autonomous(name="Auto Blue (Observation Zone)", group="Autons")
+public class AutoBlueObservation extends LinearOpMode{
 
     RobotSystem robot;
     Synchronizer synchronizer;
 
-    public static double clawOpen = 0.2;
-    public static double clawClosed = 0.1;
+    public static double clawOpen = ClawConstants.OPEN_POSITION;
+    public static double clawClosed = ClawConstants.CLOSED_POSITION;
 
-    public static double elbowUp = 0.275;
-    public static double elbowDown = 0.53;
+    public static double elbowUp = ElbowConstants.UP_POSITION;
+    public static double elbowDown = ElbowConstants.DOWN_POSITION;
 
     @Override
     public void runOpMode() throws InterruptedException {
-        this.robot = new RobotSystem(hardwareMap, new Pose2d(-24, 60, new Rotation2d(Math.toRadians(-90))), this);
+        this.robot = new RobotSystem(hardwareMap, new Pose2d(-24, 60, new Rotation2d(Math.toRadians(-90))), false, this);
         robot.inDep.setClawPosition(clawClosed);
         robot.inDep.setElbowPosition(0.3);
         initSynchronizer();
