+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/autons/AutoRedBasket.java
@@ -11,9 +11,11 @@ import org.firstinspires.ftc.teamcode.RobotSystem;
 import org.firstinspires.ftc.teamcode.opmodes.calibration.Drawing;
 import org.firstinspires.ftc.teamcode.synchropather.systems.__util__.Synchronizer;
 import org.firstinspires.ftc.teamcode.synchropather.systems.__util__.TimeSpan;
+import org.firstinspires.ftc.teamcode.synchropather.systems.claw.ClawConstants;
 import org.firstinspires.ftc.teamcode.synchropather.systems.claw.ClawPlan;
 import org.firstinspires.ftc.teamcode.synchropather.systems.claw.ClawState;
 import org.firstinspires.ftc.teamcode.synchropather.systems.claw.movements.LinearClaw;
+import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.ElbowConstants;
 import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.ElbowPlan;
 import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.ElbowState;
 import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.movements.LinearElbow;
@@ -26,23 +28,22 @@ import org.firstinspires.ftc.teamcode.synchropather.systems.rotation.movements.L
 import org.firstinspires.ftc.teamcode.synchropather.systems.translation.TranslationPlan;
 import org.firstinspires.ftc.teamcode.synchropather.systems.translation.TranslationState;
 import org.firstinspires.ftc.teamcode.synchropather.systems.translation.movements.CRSplineTranslation;
-import org.firstinspires.ftc.teamcode.synchropather.systems.translation.movements.LinearTranslation;
 
-@Autonomous(name="Auto Red Human Specimen", group="Autons")
-public class AutoRedHumanSpecimen extends LinearOpMode {
+@Autonomous(name="Auto Red (Basket)", group="Autons")
+public class AutoRedBasket extends LinearOpMode {
 
     RobotSystem robot;
     Synchronizer synchronizer;
 
-    public static double clawOpen = 0.2;
-    public static double clawClosed = 0.1;
+    public static double clawOpen = ClawConstants.OPEN_POSITION;
+    public static double clawClosed = ClawConstants.CLOSED_POSITION;
 
-    public static double elbowUp = 0.275;
-    public static double elbowDown = 0.53;
+    public static double elbowUp = ElbowConstants.UP_POSITION;
+    public static double elbowDown = ElbowConstants.DOWN_POSITION;
 
     @Override
     public void runOpMode() throws InterruptedException {
-        this.robot = new RobotSystem(hardwareMap, new Pose2d(-24, 60, new Rotation2d(Math.toRadians(-90))), this);
+        this.robot = new RobotSystem(hardwareMap, new Pose2d(-40, -60, new Rotation2d(Math.toRadians(90))), true, this);
         robot.inDep.setClawPosition(clawClosed);
         robot.inDep.setElbowPosition(0.3);
         initSynchronizer();
@@ -62,11 +63,13 @@ public class AutoRedHumanSpecimen extends LinearOpMode {
         synchronizer.stop();
     }
 
+
     private void initSynchronizer() {
+        // place preloaded specimen
         CRSplineTranslation spline1 = new CRSplineTranslation(0,
-                new TranslationState(24,-60),
-                new TranslationState(10, -45),
-                new TranslationState(10, -34)
+                new TranslationState(-40,-60),
+                new TranslationState(-10, -45),
+                new TranslationState(-10, -34)
         );
 
         LinearRotation still = new LinearRotation(0,
@@ -90,11 +93,10 @@ public class AutoRedHumanSpecimen extends LinearOpMode {
         );
 
         CRSplineTranslation splinePark = new CRSplineTranslation(liftPreload2.getEndTime(),
-                new TranslationState(10, -34),
-                new TranslationState(20, -40),
-                new TranslationState(36, -36),
-                new TranslationState(36, -12),
-                new TranslationState(24, 0)
+                new TranslationState(-10, -34),
+                new TranslationState(-36, -36),
+                new TranslationState(-36, -12),
+                new TranslationState(-24, 0)
         );
 
         TranslationPlan translationPlan = new TranslationPlan(robot,
