--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/autons/AutoRedBasketSpecimen.java
@@ -1,142 +1,0 @@
-package org.firstinspires.ftc.teamcode.opmodes.autons;
-
-import com.acmerobotics.dashboard.FtcDashboard;
-import com.acmerobotics.dashboard.telemetry.TelemetryPacket;
-import com.arcrobotics.ftclib.geometry.Pose2d;
-import com.arcrobotics.ftclib.geometry.Rotation2d;
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-import org.firstinspires.ftc.teamcode.RobotSystem;
-import org.firstinspires.ftc.teamcode.opmodes.calibration.Drawing;
-import org.firstinspires.ftc.teamcode.synchropather.systems.__util__.Synchronizer;
-import org.firstinspires.ftc.teamcode.synchropather.systems.__util__.TimeSpan;
-import org.firstinspires.ftc.teamcode.synchropather.systems.claw.ClawPlan;
-import org.firstinspires.ftc.teamcode.synchropather.systems.claw.ClawState;
-import org.firstinspires.ftc.teamcode.synchropather.systems.claw.movements.LinearClaw;
-import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.ElbowPlan;
-import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.ElbowState;
-import org.firstinspires.ftc.teamcode.synchropather.systems.elbow.movements.LinearElbow;
-import org.firstinspires.ftc.teamcode.synchropather.systems.lift.LiftPlan;
-import org.firstinspires.ftc.teamcode.synchropather.systems.lift.LiftState;
-import org.firstinspires.ftc.teamcode.synchropather.systems.lift.movements.LinearLift;
-import org.firstinspires.ftc.teamcode.synchropather.systems.rotation.RotationPlan;
-import org.firstinspires.ftc.teamcode.synchropather.systems.rotation.RotationState;
-import org.firstinspires.ftc.teamcode.synchropather.systems.rotation.movements.LinearRotation;
-import org.firstinspires.ftc.teamcode.synchropather.systems.translation.TranslationPlan;
-import org.firstinspires.ftc.teamcode.synchropather.systems.translation.TranslationState;
-import org.firstinspires.ftc.teamcode.synchropather.systems.translation.movements.CRSplineTranslation;
-import org.firstinspires.ftc.teamcode.synchropather.systems.translation.movements.LinearTranslation;
-
-@Autonomous(name="Auto Red Basket Specimen", group="Autons")
-public class AutoRedBasketSpecimen extends LinearOpMode {
-
-    RobotSystem robot;
-    Synchronizer synchronizer;
-
-    public static double clawOpen = 0.2;
-    public static double clawClosed = 0.1;
-
-    public static double elbowUp = 0.275;
-    public static double elbowDown = 0.53;
-
-    @Override
-    public void runOpMode() throws InterruptedException {
-        this.robot = new RobotSystem(hardwareMap, new Pose2d(40, 60, new Rotation2d(Math.toRadians(-90))), this);
-        robot.inDep.setClawPosition(clawClosed);
-        robot.inDep.setElbowPosition(0.3);
-        initSynchronizer();
-
-        waitForStart();
-
-        synchronizer.start();
-        while (opModeIsActive() && synchronizer.update()) {
-            robot.localization.update();
-            TelemetryPacket packet = new TelemetryPacket();
-            packet.fieldOverlay().setStroke("#3F51B5");
-            Drawing.drawRobot(packet.fieldOverlay(), robot.localization.getPose());
-            if (robot.opMode.gamepad1.triangle)
-                Drawing.drawTargetPose(packet.fieldOverlay(), new Pose2d(robot.drive.targetX, robot.drive.targetY, new Rotation2d(robot.drive.targetH)));
-            FtcDashboard.getInstance().sendTelemetryPacket(packet);
-        }
-        synchronizer.stop();
-    }
-
-
-    private void initSynchronizer() {
-        // place preloaded specimen
-        CRSplineTranslation spline1 = new CRSplineTranslation(0,
-                new TranslationState(-40,-60),
-                new TranslationState(-10, -45),
-                new TranslationState(-10, -34)
-        );
-
-        LinearRotation still = new LinearRotation(0,
-                new RotationState(Math.toRadians(90)),
-                new RotationState(Math.toRadians(90))
-        );
-
-        RotationPlan rotationPlan = new RotationPlan(robot,
-                still
-        );
-
-
-        LinearLift liftPreload1 = new LinearLift(new TimeSpan(spline1.getStartTime(), spline1.getEndTime()-0.5),
-                new LiftState(0),
-                new LiftState(1500)
-        );
-
-        LinearLift liftPreload2 = new LinearLift(spline1.getEndTime()-0.5,
-                new LiftState(1500),
-                new LiftState(0)
-        );
-
-        CRSplineTranslation splinePark = new CRSplineTranslation(liftPreload2.getEndTime(),
-                new TranslationState(-10, -34),
-                new TranslationState(-36, -36),
-                new TranslationState(-36, -12),
-                new TranslationState(-24, 0)
-        );
-
-        TranslationPlan translationPlan = new TranslationPlan(robot,
-                spline1,
-                splinePark
-        );
-
-        LiftPlan liftPlan = new LiftPlan(robot,
-                liftPreload1,
-                liftPreload2
-        );
-
-
-        // claw
-        LinearClaw claw1 = new LinearClaw(liftPreload2.getStartTime()+.66,
-                new ClawState(clawClosed),
-                new ClawState(clawOpen)
-        );
-
-
-        ClawPlan clawPlan = new ClawPlan(robot,
-                claw1
-        );
-
-        //elbow
-        LinearElbow elbowStill = new LinearElbow(claw1.getStartTime(), //goes down to sample
-                new ElbowState(elbowUp),
-                new ElbowState(elbowUp)
-        );
-
-
-        ElbowPlan elbowPlan = new ElbowPlan(robot,
-                elbowStill
-        );
-
-        this.synchronizer = new Synchronizer(
-                translationPlan,
-                rotationPlan,
-                liftPlan,
-                elbowPlan,
-                clawPlan
-        );
-    }
-}
