+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/calibration/ExampleSynchroPather/DebuggingSynchroPather.java
@@ -26,7 +26,7 @@ public class DebuggingSynchroPather extends LinearOpMode {
 
     @Override
     public void runOpMode() throws InterruptedException {
-        this.robot = new RobotSystem(hardwareMap, new Pose2d(0, 0, new Rotation2d(0)), this);
+        this.robot = new RobotSystem(hardwareMap, new Pose2d(0, 0, new Rotation2d(0)), false, this);
         initSynchronizer();
 
         waitForStart();
