+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/calibration/ExampleSynchroPather/ExampleSynchroPatherPickUpSample.java
@@ -30,7 +30,7 @@ public class ExampleSynchroPatherPickUpSample extends LinearOpMode {
 
     @Override
     public void runOpMode() throws InterruptedException {
-        this.robot = new RobotSystem(hardwareMap, new Pose2d(0, 0, new Rotation2d(0)), this);
+        this.robot = new RobotSystem(hardwareMap, new Pose2d(0, 0, new Rotation2d(0)), false, this);
         robot.inDep.setClawPosition(clawOpen);
         robot.inDep.setElbowPosition(elbowUp);
         initSynchronizers();
