+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/calibration/ExampleSynchroPather/claw/ExampleSynchroPatherClawAuto.java
@@ -34,7 +34,7 @@ public class ExampleSynchroPatherClawAuto extends LinearOpMode {
 
     @Override
     public void runOpMode() throws InterruptedException {
-        this.robot = new RobotSystem(hardwareMap, new Pose2d(0, 0, new Rotation2d(0)), this);
+        this.robot = new RobotSystem(hardwareMap, new Pose2d(0, 0, new Rotation2d(0)), false, this);
         robot.inDep.setClawPosition(clawClosed);
         initSynchronizer();
