--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/synchropather/toolkit/tuning/blah.java
@@ -1,3 +1,0 @@
-// trust this is totally a tuning opmode
-package synchropather.toolkit.tuning;
-
